Fix broken side panel from #657, revert metadata to singleton (#658)
